server/api/requirements.txt
flask==2.2.2
psycopg2-binary==2.9.5
python-dotenv==0.21.1